+++ src/pages/signup/SignUp.jsx
@@ -1,37 +1,37 @@
 import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
-import Submit from '@components/ui/button/Submit';
-import Input from '@components/ui/Input';
-import TextArea from '@components/ui/TextArea';
 import useUserApis from '@hooks/apis/useUserApis.mjs';
 import useFileApis from '@hooks/apis/useFileApis.mjs';
 import { useState } from 'react';
 import Section from '@components/ui/Section';
 import * as S from '@styles/signup/signup.style';
 import Text from '@components/ui/Text';
-import ValidationButton from '@components/ui/button/ValidationButton';
 import Toast from '@components/ui/Toast';
+import Required from '@pages/signup/Required';
+import Optional from '@pages/signup/Optional';
 
 function SignUp() {
-  const [emailMsg, setEmailMsg] = useState('');
   const navigate = useNavigate();
-  const { postSignUp, getEmailVerify } = useUserApis();
+  const { postSignUp } = useUserApis();
   const { postSingleFile } = useFileApis();
-  const {
-    register,
-    handleSubmit,
-    watch,
-    setError,
-    formState: { errors },
-  } = useForm({ mode: 'onChange' });
-  const { email } = watch();
+  const { setError } = useForm({ mode: 'onChange' });
   const [toast, setToast] = useState({
     show: false,
     message: '',
-    type: '',
   });
+  const [formData, setFormData] = useState({});
+  const [step, setStep] = useState(0);
 
-  const onSubmit = async (formData) => {
+  const handleRequiredSubmit = (formData) => {
+    // Required 컴포넌트에서 입력된 데이터를 formData에 저장
+    setFormData((prevData) => ({ ...prevData, ...formData }));
+    setStep(1);
+    console.log(formData);
+  };
+
+  const handleOptionalSubmit = async (formData) => {
+    setFormData((prevData) => ({ ...prevData, ...formData }));
+
     try {
       formData.type = 'seller';
       console.log(formData);
@@ -57,95 +57,18 @@ function SignUp() {
     }
   };
 
-  const checkDuplicateEmail = async () => {
-    try {
-      const res = await getEmailVerify(email);
-      if (res.status === 200) {
-        setEmailMsg('회원가입 가능한 이메일입니다.');
-      }
-    } catch (err) {
-      setEmailMsg(err.response?.data.message);
-    }
-  };
-
   return (
     <Section>
+      <Text typography="display_l">회원가입</Text>
+      {toast.show && <Toast setToast={setToast} text={toast.message} />}
+      {/* <S.SignUpForm onSubmit={handleSubmit(onSubmit)}> */}
       <S.SignUpWrapper>
-        <Text typography="display_l">회원가입</Text>
-        {toast.show && <Toast setToast={setToast} text={toast.message} />}
-
-        <S.SignUpForm onSubmit={handleSubmit(onSubmit)}>
-          <S.SignUpInputWrapper>
-            <div>
-              <Input
-                type="text"
-                id="name"
-                label="이름"
-                height="40px"
-                placeholder="이름을 입력하세요"
-                error={errors.name && errors.name.message}
-                {...register('name', {
-                  required: '이름을 입력하세요.',
-                  minLength: {
-                    value: 2,
-                    message: '이름을 2글자 이상 입력하세요.',
-                  },
-                })}
-              />
-            </div>
-
-            <S.SignUpEmailWrapper>
-              <div>
-                <Input
-                  type="email"
-                  id="email"
-                  label="이메일"
-                  placeholder="이메일을 입력하세요"
-                  error={errors.email && errors.email.message}
-                  {...register('email', {
-                    required: '이메일을 입력하세요.',
-                    pattern: {
-                      value: /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/,
-                      message: '이메일 형식이 아닙니다.',
-                    },
-                  })}
-                />
-              </div>
-              <S.SignUpValidation>
-                {emailMsg && (
-                  <Text typography="bold_s" color="primary02">
-                    {emailMsg}
-                  </Text>
-                )}
-                <ValidationButton type="button" disabled={email === '' || errors.email ? true : false} onClick={checkDuplicateEmail}>
-                  중복 확인
-                </ValidationButton>
-              </S.SignUpValidation>
-            </S.SignUpEmailWrapper>
-            <div>
-              <Input
-                type="password"
-                id="password"
-                label="비밀번호"
-                placeholder="비밀번호를 입력하세요"
-                error={errors.password && errors.password.message}
-                {...register('password', {
-                  required: '비밀번호를 입력하세요.',
-                })}
-              />
-            </div>
-            <div>
-              <Input type="file" accept="image/*" id="profileImage" label="프로필 이미지" placeholder="이미지를 선택하세요" {...register('profileImage')} />
-            </div>
-            <div>
-              <TextArea label="자기소개 (40자 이내)" type="txt" id="introduction" placeholder="자기소개를 입력해주세요" rows="5" {...register('introduction')} />
-            </div>
-          </S.SignUpInputWrapper>
-          <div>
-            <Submit>회원가입</Submit>
-          </div>
-        </S.SignUpForm>
+        <S.SignUpInputWrapper>
+          {step === 0 && <Required setStep={setStep} onSubmit={handleRequiredSubmit} />}
+          {step === 1 && <Optional setStep={setStep} onSubmit={handleOptionalSubmit} />}
+        </S.SignUpInputWrapper>
       </S.SignUpWrapper>
+      {/* </S.SignUpForm> */}
     </Section>
   );
 }
